Split threads into 2 services
