+++ scripts/globusmonitor/globus_monitor_service.py
@@ -5,7 +5,7 @@
     and start up an API to listen on the specified port for new
     Globus task IDs. It will then monitor the specified Globus
     directory and query the Globus API until that task ID has
-    succeeded or failed, and notify Clowder accordingly.
+    succeeded or failed, and update Postgres accordingly.
 """
 
 import os, shutil, json, time, datetime, thread, copy, atexit, collections, fcntl
@@ -74,17 +74,6 @@ config = {}
 }, {...}, {...}, ...}"""
 activeTasks = {}
 
-"""List of Globus IDs that need to be moved into Clowder; details stored in /completed folder"""
-unprocessedTasks = []
-
-"""Maps dataset/collection name to clowder UUID
-{
-    "name": "UUID"
-}
-"""
-datasetMap = {}
-collectionMap = {}
-
 app = Flask(__name__)
 api = restful.Api(app)
 
@@ -145,6 +134,7 @@ def updateNestedDict(existing, new):
 """Return small JSON object with information about monitor health"""
 def getStatus():
     activeTaskCount = len(activeTasks)
+    unprocessedTasks = readTasksByStatus("SUCCEEDED", True)
 
     return {
         "active_task_count": activeTaskCount,
@@ -162,132 +152,6 @@ def loadJsonFile(filename):
     except IOError:
         logger.error("- unable to open %s" % filename)
         return {}
-
-"""Find dataset id if dataset exists, creating if necessary"""
-def fetchDatasetByName(datasetName, requestsSession, spaceOverrideId=None):
-    if datasetName not in datasetMap:
-        # Fetch collection & space IDs (creating collections if necessary) to post with the new dataset
-        if datasetName.find(" - ") > -1:
-            sensorName = datasetName.split(" - ")[0]
-            timestamp = datasetName.split(" - ")[1].split("__")[0]
-            sensCollId = fetchCollectionByName(sensorName, requestsSession)
-            timeCollId = fetchCollectionByName(timestamp, requestsSession)
-        if not spaceOverrideId and config['clowder']['primary_space'] != "":
-            spaceOverrideId = config['clowder']['primary_space']
-        spaceId = ', "space": ["%s"]' % spaceOverrideId if spaceOverrideId else ''
-
-        dataObj = '{"name": "%s", "collection": ["%s","%s"]%s}' % (datasetName, sensCollId, timeCollId, spaceId)
-        ds = requestsSession.post(config['clowder']['host']+"/api/datasets/createempty",
-                                  headers={"Content-Type": "application/json"},
-                                  data=dataObj)
-
-        if ds.status_code == 200:
-            dsid = ds.json()['id']
-            datasetMap[datasetName] = dsid
-            writeDatasetRecordToDatabase(datasetName, dsid)
-            logger.info("++ created dataset %s (%s)" % (datasetName, dsid), extra={
-                "dataset_id": dsid,
-                "dataset_name": datasetName,
-                "action": "DATASET CREATED"
-            })
-            return dsid
-        else:
-            logger.error("- cannot create dataset (%s: %s)" % (ds.status_code, ds.text))
-            return None
-
-    else:
-        # We have a record of it, but check that it still exists before returning the ID
-        dsid = datasetMap[datasetName]
-        return dsid
-
-        # TODO: re-enable this check once backlog is caught up
-        ds = requestsSession.get(config['clowder']['host']+"/api/datasets/"+dsid)
-        if ds.status_code == 200:
-            logger.info("- dataset %s already exists (%s)" % (datasetName, dsid))
-            return dsid
-        else:
-            # Query the database just in case, before giving up and creating a new dataset
-            dstitlequery = requestsSession.get(config['clowder']['host']+"/api/datasets?title="+datasetName)
-            if dstitlequery.status_code == 200:
-                results = dstitlequery.json()
-                if len(results) > 0:
-                    return results[0]['id']
-                else:
-                    logger.error("- cannot find dataset %s; creating new dataset %s" % (dsid, datasetName))
-                    # Could not find dataset so we'll just delete the record and create a new one
-                    del datasetMap[datasetName]
-                    return fetchDatasetByName(datasetName, requestsSession, spaceOverride)
-            else:
-                logger.error("- cannot find dataset %s; creating new dataset %s" % (dsid, datasetName))
-                # Could not find dataset so we'll just delete the record and create a new one
-                del datasetMap[datasetName]
-                return fetchDatasetByName(datasetName, requestsSession, spaceOverride)
-
-"""Find list of file objects in a given dataset"""
-def fetchDatasetFileList(datasetId, requestsSession):
-    clowkey = config['clowder']['secret_key']
-    filelist = requests.get(config['clowder']['host']+"/api/datasets/%s/listFiles?key=%s" % (datasetId, clowkey),
-                              headers={"Content-Type": "application/json"})
-
-    if filelist.status_code == 200:
-        return filelist.json()
-    else:
-        logger.error("- cannot find file list for dataset %s" % datasetId)
-        return []
-
-"""Find dataset id if dataset exists, creating if necessary"""
-def fetchCollectionByName(collectionName, requestsSession):
-    if collectionName not in collectionMap:
-        coll = requestsSession.post(config['clowder']['host']+"/api/collections",
-                                    headers={"Content-Type": "application/json"},
-                                    data='{"name": "%s", "description": ""}' % collectionName)
-        time.sleep(1)
-
-        if coll.status_code == 200:
-            collid = coll.json()['id']
-            collectionMap[collectionName] = collid
-            writeCollectionRecordToDatabase(collectionName, collid)
-            logger.info("++ created collection %s (%s)" % (collectionName, collid), extra={
-                "collection_id": collid,
-                "collection_name": collectionName,
-                "action": "CREATED COLLECTION"
-            })
-            # Add new collection to primary space if defined
-            if config['clowder']['primary_space'] != "":
-                requestsSession.post(config['clowder']['host']+"/api/spaces/%s/addCollectionToSpace/%s" %
-                                     (config['clowder']['primary_space'], collid))
-            return collid
-        else:
-            logger.error("- cannot create collection (%s: %s)" % (coll.status_code, coll.text))
-            return None
-
-    else:
-        # We have a record of it, but check that it still exists before returning the ID
-        collid = collectionMap[collectionName]
-        return collid
-
-        # TODO: re-enable this check once backlog is caught up
-        coll = requestsSession.get(config['clowder']['host']+"/api/collections/"+collid)
-        if coll.status_code == 200:
-            logger.info("- collection %s already exists (%s)" % (collectionName, collid))
-            return collid
-        else:
-            # Query the database just in case, before giving up and creating a new dataset
-            colltitlequery = requestsSession.get(config['clowder']['host']+"/api/collections?title="+collectionName)
-            if colltitlequery.status_code == 200:
-                results = colltitlequery.json()
-                if len(results) > 0:
-                    return results[0]['id']
-                else:
-                    logger.error("- cannot find collection %s; creating new collection %s" % (collid, collectionName))
-                    # Could not find collection so we'll just delete the record and create a new one
-                    del collectionMap[collectionName]
-                    return fetchCollectionByName(collectionName, requestsSession)
-            else:
-                logger.error("- cannot find collection %s; creating new collection %s" % (collid, collectionName))
-                # Could not find collection so we'll just delete the record and create a new one
-                del collectionMap[collectionName]
-                return fetchCollectionByName(collectionName, requestsSession)
 
 
 # ----------------------------------------------------------
@@ -429,23 +293,6 @@ def readTasksByStatus(status, id_only=False):
     curs.close()
 
     return results
-
-"""Fetch mappings of dataset/collection name to Clowder ID"""
-def readRecordsFromDatabase():
-    q_fetch_datas = "SELECT * FROM datasets;"
-    q_detch_colls = "SELECT * FROM collections;"
-
-    curs = psql_conn.cursor()
-    logger.debug("Fetching dataset mappings from PostgreSQL...")
-    curs.execute(q_fetch_datas)
-    for currds in curs:
-        datasetMap[currds[0]] = currds[1]
-
-    logger.debug("Fetching collection mappings from PostgreSQL...")
-    curs.execute(q_detch_colls)
-    for currco in curs:
-        collectionMap[currco[0]] = currco[1]
-    curs.close()
 
 """Write dataset (name -> clowder_id) mapping to PostgreSQL database"""
 def writeDatasetRecordToDatabase(dataset_name, dataset_id):
@@ -593,42 +440,7 @@ class GlobusTask(restful.Resource):
             del activeTasks[task['globus_id']]
             return 204
 
-""" /metadata
-POST metadata for a Clowder dataset without requiring Globus task pipeline"""
-class MetadataLoader(restful.Resource):
-
-    @requires_auth
-    def post(self):
-        req = request.get_json(force=True)
-        globUser = req['user']
-        clowderUser = config['clowder']['user_map'][globUser]['clowder_user']
-        clowderPass = config['clowder']['user_map'][globUser]['clowder_pass']
-
-        sess = requests.Session()
-        sess.auth = (clowderUser, clowderPass)
-
-        md = clean_json_keys(req['md'])
-        spaceoverride = req['space_name'] if 'space_name' in req else None
-
-        dsid = fetchDatasetByName(req['dataset'], sess, spaceoverride)
-        if dsid:
-            logger.info("- adding metadata to dataset "+dsid, extra={
-                "dataset_id": dsid,
-                "action": "METADATA ADDED",
-                "metadata": md
-            })
-            dsmd = sess.post(config['clowder']['host']+"/api/datasets/"+dsid+"/metadata",
-                             headers={'Content-Type':'application/json'},
-                             data=json.dumps(md))
-
-            if dsmd.status_code != 200:
-                logger.error("- cannot add dataset metadata (%s: %s)" % (dsmd.status_code, dsmd.text))
-
-            return dsmd.status_code
-        else:
-            return "Dataset could not be accessed", 500
-
-""" / status
+""" /status
 Return basic information about monitor for health checking"""
 class MonitorStatus(restful.Resource):
 
@@ -637,7 +449,6 @@ class MonitorStatus(restful.Resource):
 
 api.add_resource(GlobusMonitor, '/tasks')
 api.add_resource(GlobusTask, '/tasks/<string:globusID>')
-api.add_resource(MetadataLoader, '/metadata')
 api.add_resource(MonitorStatus, '/status')
 
 
@@ -676,152 +487,8 @@ def getGlobusStatus(task):
     else:
         return None
 
-"""Send Clowder necessary details to load local file after Globus transfer complete"""
-def notifyClowderOfCompletedTask(task):
-    # Verify that globus user has a mapping to clowder credentials in config file
-    globUser = task['user']
-    userMap = config['clowder']['user_map']
-
-    if globUser in userMap:
-        logger.info("%s task complete; notifying Clowder" % task['globus_id'], extra={
-            "globus_id": task['globus_id'],
-            "action": "NOTIFYING CLOWDER OF COMPLETION"
-        })
-        clowderHost = config['clowder']['host']
-        clowderUser = userMap[globUser]['clowder_user']
-        clowderPass = userMap[globUser]['clowder_pass']
-
-        sess = requests.Session()
-        sess.auth = (clowderUser, clowderPass)
-
-        # This will be false if any files in the task have errors; task will be revisited
-        allDone = True
-
-        # Prepare upload object with all file(s) found
-        updatedTask = safeCopy(task)
-
-        spaceoverride = task['contents']['space_id'] if 'space_id' in task['contents'] else None
-        for ds in task['contents']:
-            filesQueued = []
-            fileFormData = []
-            datasetMD = None
-            datasetMDFile = False
-            lastFile = None
-
-            # Assign dataset-level metadata if provided
-            if "md" in task['contents'][ds]:
-                datasetMD = clean_json_keys(task['contents'][ds]['md'])
-
-            # Add local files to dataset by path
-            if 'files' in task['contents'][ds]:
-                for f in task['contents'][ds]['files']:
-                    fobj = task['contents'][ds]['files'][f]
-                    if 'clowder_id' not in fobj or fobj['clowder_id'] == "" or fobj['clowder_id'] == "FILE NOT FOUND":
-                        if os.path.exists(fobj['path']):
-                            if f.find("metadata.json") == -1:
-                                if 'md' in fobj:
-                                    # Use [1,-1] to avoid json.dumps wrapping quotes
-                                    # Replace \" with " to avoid json.dumps escaping quotes
-                                    mdstr = ', "md":' + json.dumps(fobj['md'])[1:-1].replace('\\"', '"')
-                                else:
-                                    mdstr = ""
-                                filesQueued.append((fobj['path'], mdstr))
-                                lastFile = f
-                            else:
-                                datasetMD = clean_json_keys(loadJsonFile(fobj['path']))
-                                datasetMDFile = f
-                        else:
-                            logger.info("%s dataset %s lists nonexistent file: %s" % (task['globus_id'], ds, fobj['path']))
-                            updatedTask['contents'][ds]['files'][fobj['name']]['clowder_id'] = "FILE NOT FOUND"
-                            writeTaskToDatabase(updatedTask)
-
-            if len(filesQueued)>0 or datasetMD:
-                dsid = fetchDatasetByName(ds, sess, spaceoverride)
-                dsFileList = fetchDatasetFileList(dsid, sess)
-                if dsid:
-                    # Only send files not already present in dataset by path
-                    for queued in filesQueued:
-                        alreadyStored = False
-                        for storedFile in dsFileList:
-                            if queued[0] == storedFile['filepath']:
-                                logger.info("- skipping file %s (already uploaded)" % queued[0])
-                                alreadyStored = True
-                                break
-                        if not alreadyStored:
-                            fileFormData.append(("file",'{"path":"%s"%s}' % (queued[0], queued[1])))
-
-                    if datasetMD:
-                        # Upload metadata
-                        dsmd = sess.post(clowderHost+"/api/datasets/"+dsid+"/metadata",
-                                         headers={'Content-Type':'application/json'},
-                                         data=json.dumps(datasetMD))
-
-                        if dsmd.status_code != 200:
-                            logger.error("- cannot add dataset metadata (%s: %s)" % (dsmd.status_code, dsmd.text))
-                            return False
-                        else:
-                            if datasetMDFile:
-                                logger.info("++ added metadata from .json file to dataset %s" % ds, extra={
-                                    "dataset_name": ds,
-                                    "dataset_id": dsid,
-                                    "action": "METADATA ADDED",
-                                    "metadata": datasetMD
-                                })
-                                updatedTask['contents'][ds]['files'][datasetMDFile]['metadata_loaded'] = True
-                                updatedTask['contents'][ds]['files'][datasetMDFile]['clowder_id'] = "attached to dataset"
-                                writeTaskToDatabase(updatedTask)
-                            else:
-                                # Remove metadata from activeTasks on success even if file upload fails in next step, so we don't repeat md
-                                logger.info("++ added metadata to dataset %s" % ds, extra={
-                                    "dataset_name": ds,
-                                    "dataset_id": dsid,
-                                    "action": "METADATA ADDED",
-                                    "metadata": datasetMD
-                                })
-                                del updatedTask['contents'][ds]['md']
-                                writeTaskToDatabase(updatedTask)
-
-                    if len(fileFormData)>0:
-                        # Upload collected files for this dataset
-                        # Boundary encoding from http://stackoverflow.com/questions/17982741/python-using-reuests-library-for-multipart-form-data
-                        logger.info("%s uploading unprocessed files belonging to %s" % (task['globus_id'], ds), extra={
-                            "dataset_id": dsid,
-                            "dataset_name": ds,
-                            "action": "UPLOADING FILES",
-                            "filelist": fileFormData
-                        })
-
-                        (content, header) = encode_multipart_formdata(fileFormData)
-                        fi = sess.post(clowderHost+"/api/uploadToDataset/"+dsid,
-                                       headers={'Content-Type':header},
-                                       data=content)
-
-                        if fi.status_code != 200:
-                            logger.error("- cannot upload files (%s - %s)" % (fi.status_code, fi.text))
-                            return False
-                        else:
-                            loaded = fi.json()
-                            if 'ids' in loaded:
-                                for fobj in loaded['ids']:
-                                    logger.info("++ added file %s" % fobj['name'])
-                                    updatedTask['contents'][ds]['files'][fobj['name']]['clowder_id'] = fobj['id']
-                                    writeTaskToDatabase(updatedTask)
-                            else:
-                                logger.info("++ added file %s" % lastFile)
-                                updatedTask['contents'][ds]['files'][lastFile]['clowder_id'] = loaded['id']
-                                writeTaskToDatabase(updatedTask)
-                else:
-                    logger.error("- dataset id for %s could not be found/created" % ds)
-                    allDone = False
-        return allDone
-    else:
-        logger.error("- cannot find clowder user credentials for Globus user %s" % globUser)
-        return False
-
 """Continually check Globus API for task updates"""
 def globusMonitorLoop():
-    global unprocessedTasks
-
     authWait = 0
     globWait = 0
     while True:
@@ -854,10 +521,6 @@ def globusMonitorLoop():
                                 fobj = task['contents'][ds]['files'][f]
                                 fobj['path'] = os.path.join(config['globus']['incoming_files_path'], fobj['path'])
 
-                    # Notify Clowder to process file if transfer successful
-                    if globusStatus == "SUCCEEDED":
-                        unprocessedTasks.append(globusID)
-
                     writeTaskToDatabase(task)
                     del activeTasks[globusID]
 
@@ -871,39 +534,6 @@ def globusMonitorLoop():
             generateAuthTokens()
             authWait = 0
 
-"""Work on completed Globus transfers to process them into Clowder"""
-def clowderSubmissionLoop():
-    global unprocessedTasks
-
-    clowderWait = config['clowder']['globus_processing_frequency'] - 1
-    while True:
-        time.sleep(1)
-        clowderWait += 1
-
-        # Check with Globus for any status updates on monitored tasks
-        if clowderWait >= config['clowder']['globus_processing_frequency']:
-            logger.info("- checking unprocessed tasks")
-            toHandle = safeCopy(unprocessedTasks)
-
-            for globusID in toHandle:
-                task = readTaskFromDatabase(globusID)
-                if task:
-                    clowderDone = notifyClowderOfCompletedTask(task)
-                    if clowderDone:
-                        logger.info("%s task successfully processed!" % globusID, extra={
-                            "globus_id": globusID,
-                            "action": "PROCESSING COMPLETE"
-                        })
-                        unprocessedTasks.remove(globusID)
-                        task['status'] = 'PROCESSED'
-                        writeTaskToDatabase(task)
-                    else:
-                        logger.error("%s not successfully sent" % globusID)
-                else:
-                    logger.error("%s unprocessed task record not found in completed tasks" % globusID)
-
-            clowderWait = 0
-
 if __name__ == '__main__':
     # Try to load custom config file, falling back to default values where not overridden
     config = loadJsonFile(os.path.join(rootPath, "config_default.json"))
@@ -916,22 +546,18 @@ if __name__ == '__main__':
     # Initialize logger handlers
     with open(os.path.join(rootPath,"config_logging.json"), 'r') as f:
         log_config = json.load(f)
-        log_config['handlers']['file']['filename'] = config["log_path"]
+        log_config['handlers']['file']['filename'] = config["monitor_log_path"]
         logging.config.dictConfig(log_config)
     logger = logging.getLogger('gantry')
 
     psql_conn = connectToPostgres()
-    readRecordsFromDatabase()
     activeTasks = readTasksByStatus("IN PROGRESS")
-    unprocessedTasks = readTasksByStatus("SUCCEEDED", True)
     generateAuthTokens()
 
-    logger.info("- initializing services")
+    logger.info("- initializing service")
     # Create thread for service to begin monitoring
     thread.start_new_thread(globusMonitorLoop, ())
     logger.info("*** Service now monitoring Globus tasks ***")
-    thread.start_new_thread(clowderSubmissionLoop, ())
-    logger.info("*** Service now waiting to process completed tasks into Clowder ***")
 
     # Create thread for API to begin listening - requires valid Globus user/pass
     apiPort = os.getenv('MONITOR_API_PORT', config['api']['port'])
